+++ src/core/library/track/LibraryTrack.cpp
@@ -174,7 +174,7 @@ bool LibraryTrack::Load(Track *target, db::Connection &db) {
         "FROM tracks t, paths p, albums al " \
         "WHERE t.id=? AND t.album_id=al.id", db);
 
-    trackQuery.BindInt(0, target->GetId());
+    trackQuery.BindInt(0, (uint64) target->GetId());
     if (trackQuery.Step() == db::Row) {
         target->SetValue("track", trackQuery.ColumnText(0));
         target->SetValue("disc", trackQuery.ColumnText(1));
@@ -192,17 +192,17 @@ bool LibraryTrack::Load(Track *target, db::Connection &db) {
         target->SetValue("album_artist_id", trackQuery.ColumnText(13));
         target->SetValue("album_id", trackQuery.ColumnText(14));
 
-        genresQuery.BindInt(0, target->GetId());
+        genresQuery.BindInt(0, (uint64) target->GetId());
         while (genresQuery.Step() == db::Row) {
             target->SetValue("genre", genresQuery.ColumnText(0));
         }
 
-        artistsQuery.BindInt(0, target->GetId());
+        artistsQuery.BindInt(0, (uint64) target->GetId());
         while (artistsQuery.Step() == db::Row) {
             target->SetValue("artist", artistsQuery.ColumnText(0));
         }
 
-        allMetadataQuery.BindInt(0, target->GetId());
+        allMetadataQuery.BindInt(0, (uint64) target->GetId());
         while (allMetadataQuery.Step() == db::Row) {
             target->SetValue(allMetadataQuery.ColumnText(1), allMetadataQuery.ColumnText(0));
         }
